Implement functional security (#23)
Refactor security for laminar-oapi to use a return type instead of throwing exceptions. This means we need to catch them for validation tools like jsonwebotken that work with throwing exceptions. But build much more complex tools is now a lot easier.

Additionally make all the examples tested with unit tests.