Yarn 2 (#46)
Switch to using yarn 2 (berry)

Yarn 2 offers a much faster and more robust experience.

Uncovered numerous problems with all of the workspace packages, as yarn 1 will happily ignore the individual package.json dependencies and just use the global node_modules folder. Yarn 2 doesn't allow you to pull stuff like this, and you need to be strict and explicit about all of your dependencies, which I think is great.

Additionally its just plain faster on all operations, sometimes ridiculously so. The drawback is that the environment needs to be carefully configured so that all calls to node pass through yarn's sdk first, updating its resolution algorithms.

The additional speed allowed for a matrix of different node versions in testing, as well as fixing a couple of small bugs that were found because of that.